Add Method to check if output has valid shape for perturbations per eval. (#1666)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/captum/pull/1666

This diff adds a new method to the `captum.attr._core.feature_ablation` module to check if the output shape of the forward function scales correctly with the input batch size when perturbations are applied. The method takes in the inputs, the number of examples, the initial evaluation, the modified evaluation, and the number of perturbations per evaluation as arguments. It then validates that the output shape of the forward function scales correctly with the input batch size.

Reviewed By: jjuncho

Differential Revision: D86976520

fbshipit-source-id: 5cb40721b835d44f023cb26d0b6939142f2122e6