Make _should_skip_inputs_and_warn a free function (#1667)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/captum/pull/1667

This diff makes a change to the `_should_skip_inputs_and_warn` function in the `captum/attr/_core/feature_ablation.py` file. The function is now a free function instead of being a method of a class. The function checks two conditions that would cause a feature group to be skipped during attribution computation: 1. If `min_examples_per_batch_grouped` is specified and any input tensor in the feature group has a batch size (0th dimension) smaller than this threshold. 2. If all input tensors in the feature group are empty

Reviewed By: sarahtranfb

Differential Revision: D87300652

fbshipit-source-id: 92b77da5e6521657897e54951a9fee910608e23d