Remove the support of skip_tokens in llm_attr target (#1685)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/captum/pull/1685

The `skip_token` was originally introduced in D62775617 . this can be further improved.

This diff reverts it.

Transformer is auto-regression model, so the sequence of token cannot be changed. It doesn't make sense to support filter tokens within the sequence.

The correct behavior is to only trim leading extra special tokens (if any) when `target` is tokenized alone (w/o input)

Reviewed By: yeyingxiao

Differential Revision: D89682911

fbshipit-source-id: 4a80b40dc920f12990c69e43e0bf21f19bc28f6f